requirements.txt
cassandra-driver
nose
numpy
pyyaml
scikit-learn
scipy
